ready to present

+++ app/Http/Controllers/UserController.php
@@ -12,24 +12,29 @@ class UserController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Http\Response
      */
     public function index()
     {
         $roles = Role::all()->pluck('name');
         $users = User::withTrashed()->paginate(10);
-        return view('admin.user.index',compact('roles', 'users'));
+        return view('admin.user.index', compact('roles', 'users'));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\User  $user
+     * @param \App\Models\User $user
      * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response
      */
     public function update(User $user, Request $request)
     {
-        $user->assignRole($request->role);
+        if ($user->id === 1)
+            return back()->with('user', $user)
+                ->with('role', 'This user`s role can`t change');
+
+        if (auth()->user()->hasRole(['super_admin', 'admin']))
+            $user->syncRoles($request->role);
         return back()->with('user', $user)
             ->with('role', 'Role assigned');
     }
@@ -37,11 +42,16 @@ class UserController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\User  $user
+     * @param \App\Models\User $user
      * @return \Illuminate\Http\RedirectResponse|\Illuminate\Http\Response
      */
     public function destroy(User $user)
     {
+        if ($user->id === 1)
+            return back()->with('user', $user)
+                ->with('delete', 'This user can`t being deactivated');
+
+
         User::find($user->id)->delete();
 
         return back()->with('user', $user)
